refactor: RoomPager에서 RoomRecycler로 일부 역할 이동
