+++ android/app/src/main/java/com/digginroom/digginroom/views/customview/roompager/RoomPager.kt
@@ -1,12 +1,10 @@
 package com.digginroom.digginroom.views.customview.roompager
 
 import android.content.Context
-import android.graphics.Point
 import android.util.AttributeSet
 import android.view.MotionEvent
 import android.widget.FrameLayout
 import android.widget.LinearLayout
-import com.digginroom.digginroom.views.customview.roomview.RoomPlayer
 import com.digginroom.digginroom.views.model.RoomModel
 
 class RoomPager(
@@ -18,11 +16,6 @@ class RoomPager(
     private val horizontalScrollPager: HorizontalScrollPager = HorizontalScrollPager(context)
     private val roomRecycler: RoomRecycler = RoomRecycler(context, GRID_SIZE)
 
-    private var targetRooms: List<Point> = listOf(Point(0, 0), Point(0, 1), Point(1, 0))
-    private var rooms: List<RoomModel> = emptyList()
-    private var currentRoomPosition = 0
-    var onNextRoom: () -> Unit = { }
-
     init {
         initVerticalScrollView()
         initHorizontalScrollView()
@@ -30,8 +23,7 @@ class RoomPager(
     }
 
     fun updateData(rooms: List<RoomModel>) {
-        this.rooms = rooms
-        navigateRooms()
+        roomRecycler.updateData(rooms)
     }
 
     private fun initVerticalScrollView() {
@@ -77,33 +69,21 @@ class RoomPager(
             if (event.action != MotionEvent.ACTION_UP) return@setOnTouchListener
             when (scrollPager.pagingState) {
                 PagingState.PREVIOUS -> {
-                    currentRoomPosition--
+                    roomRecycler.currentRoomPosition--
                     scrollPager.scrollPosition--
                 }
 
                 PagingState.CURRENT -> {}
                 PagingState.NEXT -> {
-                    currentRoomPosition++
+                    roomRecycler.currentRoomPosition++
                     scrollPager.scrollPosition++
                 }
             }
 
-            if (currentRoomPosition <= 0) {
-                currentRoomPosition = 0
+            if (roomRecycler.currentRoomPosition <= 0) {
+                roomRecycler.currentRoomPosition = 0
             } else if (scrollPager.scrollPosition <= 0) {
-                repeat(GRID_SIZE) {
-                    val child = roomRecycler.getChildAt(
-                        scrollPager.calculateEndChildPosition(
-                            it,
-                            GRID_SIZE
-                        )
-                    )
-                    roomRecycler.removeView(child)
-                    roomRecycler.addView(
-                        child,
-                        scrollPager.calculateStartChildPosition(it, GRID_SIZE)
-                    )
-                }
+                roomRecycler.recycleView(scrollPager)
 
                 scrollPager.scrollPosition++
 
@@ -111,20 +91,8 @@ class RoomPager(
                     scrollPager.screenSize
                 )
             } else if (scrollPager.scrollPosition >= GRID_SIZE - 1) {
-                onNextRoom()
-                repeat(GRID_SIZE) {
-                    val child = roomRecycler.getChildAt(
-                        scrollPager.calculateStartChildPosition(
-                            it,
-                            GRID_SIZE
-                        )
-                    )
-                    roomRecycler.removeView(child)
-                    roomRecycler.addView(
-                        child,
-                        scrollPager.calculateEndChildPosition(it, GRID_SIZE)
-                    )
-                }
+                roomRecycler.onNextRoom()
+                roomRecycler.recycleView(scrollPager)
 
                 scrollPager.scrollPosition--
 
@@ -133,12 +101,6 @@ class RoomPager(
                 )
             }
 
-            targetRooms = listOf(
-                Point(scrollPager.scrollPosition - 1, scrollPager.scrollPosition),
-                Point(scrollPager.scrollPosition, scrollPager.scrollPosition),
-                Point(scrollPager.scrollPosition + 1, scrollPager.scrollPosition)
-            )
-
             roomRecycler.playCurrentRoomPlayer(verticalScrollPager.scrollPosition * GRID_SIZE + horizontalScrollPager.scrollPosition)
             scrollPager.post {
                 scrollPager.smoothScrollTo(
@@ -146,21 +108,6 @@ class RoomPager(
                 )
             }
         }
-    }
-
-    private fun navigateRooms() {
-        targetRooms.forEachIndexed { index, point ->
-            val recyclePosition = currentRoomPosition + index - 1
-            if (recyclePosition in rooms.indices) {
-                (roomRecycler.getChildAt(pointToScalar(point)) as RoomPlayer).navigate(
-                    rooms[recyclePosition]
-                )
-            }
-        }
-    }
-
-    private fun pointToScalar(point: Point): Int {
-        return point.y * GRID_SIZE + point.x
     }
 
     companion object {
